+++ src/Cryptol/ModuleSystem/NamingEnv.hs
@@ -6,14 +6,14 @@
 -- Stability   :  provisional
 -- Portability :  portable
 
-{-# LANGUAGE DeriveAnyClass #-}
-{-# LANGUAGE DeriveGeneric #-}
 {-# LANGUAGE BlockArguments #-}
-{-# LANGUAGE OverloadedStrings #-}
 {-# LANGUAGE RecordWildCards #-}
 -- See Note [-Wincomplete-uni-patterns and irrefutable patterns] in Cryptol.TypeCheck.TypePat
 {-# OPTIONS_GHC -Wno-incomplete-uni-patterns #-}
-module Cryptol.ModuleSystem.NamingEnv where
+module Cryptol.ModuleSystem.NamingEnv
+  ( module Cryptol.ModuleSystem.NamingEnv.Types
+  , module Cryptol.ModuleSystem.NamingEnv
+  ) where
 
 import Data.Maybe (mapMaybe,maybeToList)
 import           Data.Map.Strict (Map)
@@ -21,9 +21,6 @@ import qualified Data.Map.Strict as Map
 import           Data.Set (Set)
 import qualified Data.Set as Set
 import           Data.Foldable(foldl')
-
-import GHC.Generics (Generic)
-import Control.DeepSeq(NFData)
 
 import Cryptol.Utils.PP
 import Cryptol.Utils.Panic (panic)
@@ -34,24 +31,7 @@ import Cryptol.ModuleSystem.Name
 import Cryptol.ModuleSystem.Names
 import Cryptol.ModuleSystem.Interface
 
-
--- | The 'NamingEnv' is used by the renamer to determine what
--- identifiers refer to.
-newtype NamingEnv = NamingEnv (Map Namespace (Map PName Names))
-  deriving (Show,Generic,NFData)
-
-instance Monoid NamingEnv where
-  mempty = NamingEnv Map.empty
-  {-# INLINE mempty #-}
-
-instance Semigroup NamingEnv where
-  NamingEnv l <> NamingEnv r =
-    NamingEnv (Map.unionWith (Map.unionWith (<>)) l r)
-
-instance PP NamingEnv where
-  ppPrec _ (NamingEnv mps)   = vcat $ map ppNS $ Map.toList mps
-    where ppNS (ns,xs) = nest 2 (vcat (pp ns : map ppNm (Map.toList xs)))
-          ppNm (x,as)  = pp x <+> "->" <+> commaSep (map pp (namesToList as))
+import Cryptol.ModuleSystem.NamingEnv.Types
 
 
 {- | This "joins" two naming environments by matching the text name.
@@ -91,14 +71,16 @@ namingEnvNames (NamingEnv xs) =
     Just (One x) -> Set.singleton x
     Just (Ambig as) -> as
 
--- | Get a unqualified naming environment for the given names
+-- | Get a naming environment for the given names.  The `PName`s correspond
+-- to the definition sites of the corresponding `Name`s, so typically they
+-- will be unqualified.  The exception is for names that comre from parameters,
+-- which are qualified with the relevant parameter.
 namingEnvFromNames :: Set Name -> NamingEnv
 namingEnvFromNames xs = NamingEnv (foldl' add mempty xs)
   where
   add mp x = let ns = nameNamespace x
-                 txt = nameIdent x
              in Map.insertWith (Map.unionWith (<>))
-                               ns (Map.singleton (mkUnqual txt) (One x))
+                               ns (Map.singleton (nameToDefPName x) (One x))
                                mp
 
 
@@ -231,8 +213,8 @@ isEmptyNamingEnv (NamingEnv mp) = Map.null mp
 
 -- | Compute an unqualified naming environment, containing the various module
 -- parameters.
-modParamsNamingEnv :: T.ModParamNames -> NamingEnv
-modParamsNamingEnv T.ModParamNames { .. } =
+modParamNamesNamingEnv :: T.ModParamNames -> NamingEnv
+modParamNamesNamingEnv T.ModParamNames { .. } =
   NamingEnv $ Map.fromList
     [ (NSValue, Map.fromList $ map fromFu $ Map.keys mpnFuns)
     , (NSType,  Map.fromList $ map fromTS (Map.elems mpnTySyn) ++
@@ -248,6 +230,11 @@ modParamsNamingEnv T.ModParamNames { .. } =
 
   fromTS ts = (toPName (T.tsName ts), One (T.tsName ts))
 
+-- | Compute a naming environment from a module parameter, qualifying it
+-- according to 'mpQual'.
+modParamNamingEnv :: T.ModParam -> NamingEnv
+modParamNamingEnv mp = maybe id qualify (T.mpQual mp) $
+  modParamNamesNamingEnv (T.mpParameters mp)
 
 -- | Generate a naming environment from a declaration interface, where none of
 -- the names are qualified.